Splitting out SuggestedPlayersSorter
